refactor: auth 도메인 enums 패키지 분리
