+++ src/main/java/com/example/WonkaoTalk/domain/auth/entity/AuthSocial.java
@@ -1,5 +1,6 @@
 package com.example.WonkaoTalk.domain.auth.entity;
 
+import com.example.WonkaoTalk.domain.auth.enums.AuthProvider;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
